chore: Reduce dependencies at their minimum of ext.core package
