+++ src/Uno.Extensions.Core/AsyncFunc.cs
