+++ src/Uno.Extensions.Core/AsyncFunc.tt
