--- src/Uno.Extensions.Core/FuncAsync.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Uno.Extensions;
-
-public delegate ValueTask<TResult> FuncAsync<TResult>(CancellationToken ct);
-public delegate ValueTask<TResult> FuncAsync<in T, TResult>(T arg, CancellationToken ct);
-public delegate ValueTask<TResult> FuncAsync<in T1, in T2, TResult>(T1 arg1, T2 arg2, CancellationToken ct);
