+++ src/Uno.Extensions.Core/Option.cs
@@ -122,42 +122,3 @@ public readonly struct Option<T> : IOption, IEquatable<Option<T>>
 			_ => $"Some({_value})",
 		};
 }
-
-//public static class OptionExtensions
-//{
-//	public static async ValueTask<Option<TResult>> MapAsync<T, TResult>(this Option<T> option, FuncAsync<T?, TResult?> projection, CancellationToken ct)
-//		=> option.IsSome(out var value) ? await projection(value, ct)
-//			: option.IsNone() ? Option<TResult>.None()
-//			: Option<TResult>.Undefined();
-//}
-
-//internal class ReferenceEqualityComparer<T> : IEqualityComparer<T>
-//{
-//	private ReferenceEqualityComparer()
-//	{
-//	}
-
-//	/// <inheritdoc />
-//	public bool Equals(T x, T y)
-//		=> object.ReferenceEquals(x, y);
-
-//	/// <inheritdoc />
-//	public int GetHashCode(T obj)
-//		=> obj?.GetHashCode() ?? 0;
-//}
-//	/// <inheritdoc />
-//	public bool Equals(T x, T y)
-//		=> object.ReferenceEquals(x, y);
-
-//	/// <inheritdoc />
-//	public int GetHashCode(T obj)
-//		=> obj?.GetHashCode() ?? 0;
-//}
-//	/// <inheritdoc />
-//	public bool Equals(T x, T y)
-//		=> object.ReferenceEquals(x, y);
-
-//	/// <inheritdoc />
-//	public int GetHashCode(T obj)
-//		=> obj?.GetHashCode() ?? 0;
-//}
