+++ src/Uno.Extensions.Core/_Compat/IsExternalInit.cs
