+++ src/Uno.Extensions.Core/_Compat/NotNullIfNotNullAttribute.cs
@@ -1,18 +1,8 @@
 ﻿using System;
+using System.Linq;
 
 #if NETSTANDARD2_0 || WINDOWS_UWP || NET461
 namespace System.Diagnostics.CodeAnalysis;
-
-[AttributeUsage(AttributeTargets.Parameter, Inherited = false)]
-internal class NotNullWhenAttribute : Attribute
-{
-	public bool ReturnValue { get; }
-
-	public NotNullWhenAttribute(bool returnValue)
-	{
-		ReturnValue = returnValue;
-	}
-}
 
 [AttributeUsage(AttributeTargets.Parameter | AttributeTargets.Property | AttributeTargets.ReturnValue, AllowMultiple = true, Inherited = false)]
 internal class NotNullIfNotNullAttribute : Attribute
