+++ src/Uno.Extensions.Reactive.Generator/Uno.Extensions.Reactive.Generator.csproj
@@ -24,12 +24,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<Compile Include="..\Uno.Extensions.Core\_Compat\**\*.cs" Link="_Compat\%(RecursiveDir)%(Filename)%(Extension)" />
 		<Compile Include="..\Uno.Extensions.Reactive\**\*Attribute.cs" Link="Attributes\%(RecursiveDir)%(Filename)%(Extension)" />
 		<Compile Include="..\Uno.Extensions.Reactive.UI\**\*Attribute.cs" Link="Attributes\%(RecursiveDir)%(Filename)%(Extension)" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <Compile Include="..\Uno.Extensions.Core\Compat\NotNullWhenAttribute.cs" Link="Utils\NotNullWhenAttribute.cs" />
 	</ItemGroup>
 
 	<ItemGroup>
