--- src/Uno.Extensions.Reactive/_Compat/IsExternalInit.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Linq;
-
-#if !NET5_0
-namespace System.Runtime.CompilerServices;
-
-internal static class IsExternalInit
-{
-}
-#endif
